+++ src/main/java/com/erebelo/springh2demo/repository/CustomerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
 
     Optional<CustomerEntity> findByEmail(String email);
